begin uncoupling program.py by moving add_tab and change_title out
(And then promptly realize I have a can I need to stop kicking a can down the road.)
