+++ hints/control/hint_notebook.py
@@ -42,30 +42,7 @@ class HintNotebook(Program):
         # Run the window
         self.root.mainloop()
 
-    def add_tab(self, tab_name: str) -> None | CTkFrame:
-        '''Create a tab in the notebook.'''
-        # If it already exists, don't bother
-        if tab_name in self.data_tabs.keys():
-            return
-
-        # Update the data tabs dict
-        self.update_data_tabs(tab_name, None)
-
-        # Find the index
-        tab_index = tab_names.data_tab_names.index(tab_name)
-
-        # Create the tab, and return it
-        return self.notebook.insert(tab_index, tab_name)
-
-    def change_title(self, seed_name: str = '') -> None:
-        '''Change the title of the window.'''
-        # The default without the seed name
-        title = 'TPR Hint Notebook'
-        # If there was a seed name, append it
-        if seed_name:
-            title = f'{title} ({seed_name})'
-
-        self.root.title(title)
+    
 
     def create_data_tabs(self) -> None:
         '''Creates the tabs that have data in their default state.'''
@@ -73,7 +50,7 @@ class HintNotebook(Program):
         # then store the notepad for later use.
         for tab_name in tab_names.data_tab_names:
             # Create the tab
-            self.add_tab(tab_name)
+            self.resetter.add_tab(tab_name)
 
             # Create the notepad that goes in it
             notepad = self.create_notepad(tab_name)
