+++ hints/utils/parse_log.py
@@ -4,6 +4,7 @@
 from hints.control.program import Program
 from hints.tabs.shopping.agitha_tab import AgithaTab
 from hints.utils.constants import folders
+from hints.utils.gui_management.reset_utils import ResetUtils
 from json import load
 from pathlib import Path
 from re import findall, sub
@@ -13,6 +14,7 @@ class ParseLog:
     '''It just, parses the spoiler log data.'''
     # The root program
     program = Program
+    resetter = ResetUtils
 
     # Spoiler log info
     spoiler_log_file = str     # The provided spoiler log
@@ -21,6 +23,7 @@ class ParseLog:
         '''Set the global var here.'''
         # Set the local program var
         self.program = program
+        self.resetter = self.program.resetter
 
     def dump_and_fill(self, spoiler_log_file: str) -> None:
         '''Take the provided path, and dump the log then fill the tabs.'''
@@ -29,7 +32,7 @@ class ParseLog:
 
         # Change the window title to include the seed name
         seed_name = findall(r'\-\-(.*?)\-\-', spoiler_log_file)[0]
-        self.program.change_title(seed_name)
+        self.resetter.change_title(seed_name)
 
         # Parse the provided data
         self.parse_spoiler_log()
